n2n: only show option to send files in advanced mode
It is too easy for new users to shoot themselves in the foot with
this, and there is the better option of uploading a file into Freenet
and then recommending it to a friend.

Uploading is better, because it makes it easy for the friend to
re-share the file with other friends (no new direct transfer is
required).