Created public and package-private interfaces for LocationManager and PeerManager.
To be able to replace the LocationManager and PeerManager with stubs
during unit tests, each class was replaced by a public interface with
all public methods of the respective class. The implementation was
moved into a class named *Impl. Furthermore, a package-private interface
extending the public interface was created for each class, as several classes
in the same package used to call package-private methods. All
references to the LocationManager and PeerManager have been
changed to the corresponding interface class.

To enable testing, the Node class has been extended with a new*-Method
for each class. These methods can be overriden by a derived class, such
that an appropriate stub is generated instead of a real instance of the
corresponding *Impl-Class.