[m3u-player] inject vanilla Javascript m3u-player into Freesites
With this change, adding a media-tag that references an m3u-list like

<audio src="stream.m3u" controls="controls">
	not supported?
    </audio>

or

<video src="stream.m3u" controls="controls">
	<a href="stream.m3u">stream.m3u</a>, starting with <a href="sff-000.ogv">sff-001.ogv</a>. Putting the video here causes freesitemgr to more likely
put the first video into the container, giving a better initial impression.
    </video>

into a Freesite will turn the media-tag into an m3u-player that
dynamically loads consecutive entries from the m3u file as needed to
provide continuous play.