User alert if datastore is below 10% of available space -- rebased without new datastore settings
Let us encourage users to increase the size of their datastore if it
is below 10% of available free disk space, which is clearly below the
40% of available free disk space we suggest in the wizard.

The wizard suggested very small sizes before, so the network can
benefit if we get nodes to upgrade their datastore sizes. It is
also possible that people have bought new larger disks and just
moved their Freenet installation.

If a user has chosen a very small size during the first time wizard,
this user alert also acts as a way for them to change their mind.

The user alert will send you to the wizard step again to reconfigure,
but will not change datastore type, just size.

The user alert is at warning level, and can be dismissed.