Make it easier to build without network access
Both the gradle distribution and all .jar dependencies can be put in
the lib directory. They can be downloaded from Freenet or from Tor
ahead of time, and the file integrity will still be validated using
the commited checksums.

I could not figure out how to get the android dependency to work
in offline mode, so let us disable it until someone does. It is not
needed when building for desktop.