Disable write local to datastore functionality when opennet enabled
The datastore is only effective at low HTL since all requests for
a block is routed towards a certain location. At high HTL we would
at best overwrite some low HTL data block with data block that is
never going to be requested from us, at worst we reveal that we or
a nearby node have requested this data block since it otherwise
would be highly unlikely for us to have it.

This feature is a security issue, while providing no meaningful
performance gain in theory. The only reason why you would want to
enable this feature is if you are bootstrapping a very small darknet,
since otherise no node would write anything to their datastores.

For nodes running opennet, force disable this option to avoid leaving
users with it enabled. For nodes running darknet, it would already
have been disabled by the security levels, so leave the value as it is.