Stop storing blocks twice
The logic was wrong, causing all blocks written to the store to also
be written to the cache. The store is written to in about half the
speed of the cache. This means effective cache storage capacity is
only 50% of what it should be, the other 50% hold copies of the store.

Whenever we get an incoming request, we already check both the cache
and store to see if we can fulfill the request. We also already make
sure the block never gets promoted from store to cache later on.

This fix may improve total network capacity significantly.