m3u-player: skip tracks that fail.
This increases the resilience against failing fragments (which do
happen in Freenet, especially in live-streaming where the problem
might actually be a failed upload).

The goal of this change is to enable listening in the background
without having to interact with the freenet player after the initial
start of playback.