Fix compression result comparison in the GzipCompressorTest for newer JDKs
Newer JDKs changed how GZIP OS(Operating System) member header value is assigned.
Test data value in the test uses fixed constant, which differs from value assigned by newer JDKs.
This caused the tests to fail on newer JDKs like 17 and 19.

The solution for this was initially described in the bug comments in:
https://freenet.mantishub.io/view.php?id=7176
This change just assigns a fixed value to the GZIP OS member header byte to the compressed data,
which becomes same as in the test data.

Also some minor test improvements are included.

As a result, project now is able to build with JDK 17 and 19


Signed-off-by: Veniamin Fernandes <veniamin.fernandes@proton.me>