Call the match function with all arguments directly.
Match is called so often, that the match-method with two arguments
took 20% of the self-time of the match-method with three arguments.

This may be inflated due to the instrumentation, though, because the
method with two arguments does nothing except for running the method
with three arguments with the default (false), so I expect less than
those 20% in wins.