Counting loop keeping the node from inserting too early was not looping
According to the comments:
We always decrement on starting a sender.
However, after that, if our HTL is above the no-cache threshold, we do not want to decrement the HTL for trivial rejections (e.g. RejectedLoop), because we would end up caching data too close to the originator. So allow 5 failures and then RNF.