RunningAverage: replace synchronization with immutable state
Simplify the implementations of the RunningAverage interface by
separating the data being kept into an immutable object. This way we no
longer have to synchronize on everything, but rather update the data
object using compare-and-set.

MedianMeanRunningAverage is not included in this effort because that
performance would be hurt due to the large amount of data kept.