RunningAverage: deprecate MedianMeanRunningAverage and replace its uses
Replace all uses of MedianMeanRunningAverage with TrivialRunningAverage
which does not eat an unbounded amount of RAM. This class was only used
for debug purposes under the logMINOR flag. Having the median value in
addition to mean value does likely not warrant the additional RAM and
CPU consumption.

Deprecate the class so that it can eventually be removed.