Remove code related to NLM load sending
This code was effectively inactive since 2013 (which is when the FIXME
in PeerNode.makeLoadStats was added: "re-enable when try NLM again").

Removing the related code not only gets rid of a few hundred lines of
code, but also slightly reduces the (huge) number of things PeerNode
needs to keep track of and simplifies the NPF logic substantially.

Note: this does not remove receiving of load stats from peer nodes.