Remove effectively unused archive restart code in ArchiveManager
The `lastSize` and `lastHash` on the ArchiveStoreContext appear to keep
track of the last observed size and hash for an archive data bucket, but
they are only ever written by the ArchiveManager when they already have
a value - which they never do.

Therefore, there are no code paths for ArchiveRestartedException to be
thrown. This has been the case for at least since 2005, and apparently
it works fine this way.

Inline the default values of these fields in ArchiveManager and prune
the now unreachable code.