Move archive entry caching out of ArchiveManager
Remove the double bookkeeping of archive entries in ArchiveManager and
ArchiveStoreContext and replace it with a central cache implementation.
This simplifies the locking (for which there were instructions in the
comments, but these were ignored in the implementation) and reduces the
number of arguments being passed around.

This also removes the "remove all cached items for an archive" logic
that would likely lead to memory leaks as not all references were
consistently released, and fixes some potential race conditions in
claiming the buckets.