Fix request distribution stats
The `count` was never incremented, and consequently the Incoming Request
Distribution and Outgoing Request Distribution panels would no longer
show up on the statistics page.

Instead of keeping track of the total count independent of the bins,
just sum the bins when we need a total count for display purposes.

While we're at it, let's get rid of the synchronization too.