🎨 Make l10n provider used in tests configurable
Both the GenericReadFilterCallback (by way of ContentFilterTest) and the Pebble
extension function test require a translation that is not the production
translation; the ContentFilterTest requires that the used translation key does
not have a translation (so the key is returned and can be verified), and the
Pebble extension function test requires a special key to verify the function
result.

So far these tests have worked because the freenet.l10n.en.properties file from
the main resources is overridden in tests by a freenet.l10n.en.properties that
serves both these tests. However, as the presence of that file breaks one of
the tests in BaseL10nTest, these tests must no longer rely on the classloader
giving them the correct data, but need to take matter into their own hands.