🐛 Fix translation properties resolution in tests
The translation properties resolution in tests is an absolute wild ride:

The build file set the current working directory for the tests to be one of the
directories that compiled files end up in, causing the tests to fail when run
from the IDE, because the IDE usually executes the tests from the project’s
root directory.

The build file sets some properties that are exclusively used in TestProperty,
where they are specified with defaults that are different from what is in the
build file, but because they are specified in the build file, these defaults
are never used.

The default values (and the overriding values) are rather nonsensical, and
incorrectly used. In BaseL10nTest, the methods createL10n() and
createTestL10n() use the values from TestProperty to define the l10n override
file mask, and to create a BaseL10n object – with completely broken values! The
first parameter of the BaseL10n constructor is not really a path, but a prefix
for a location on the _classpath_. Constructs like “../main/freenet/l10n” do
not make any sense!

The combination of these mishaps means that at least one of the tests was
subtly broken; the test “testStrings” in BaseL10nTest is supposed to verify the
syntactical correctness of all translation values, but due to the BaseL10n
object being a complete mess, the properties files were never actually loaded,
still allowing the test to pass without actually verifying anything.

So this commit fixes all of these problems: the build file was liberated from
some of the custom fiddliness, the defaults in TestProperty were set to
something sensible, the values were used correctly to create BaseL10n objects,
and the properties for tests have been moved to different filenames, allowing
tests to actually choose which set of resources they access.