🐛 Fix Atom XML being built incorrectly
This reverses the order of the entries in the XML file, but according to
every spec I can find, this should not matter, as every entry has its
own timestamp.

This also add the ElementBuilder/XmlBuilder classes for quickly
constructing XML documents that will always be valid. I would love to
move these somewhere else, but have not found a suitable location; as
long as freenet.support is being advertised as “other applications could
use this, too” I am very reluctant to put anything more in there than
there already is, because more public API is not good.