Fix noderef folder permissions for Debian package
Due to the postinst script being misnamed it does not run on install,
and the /etc/freenet/noderef ends up being owned by root:root. As a
consequence, the node cannot write in the directory, noderefs and peer
lists are not persisted, and node identity is lost on every restart.

Rename the post-installation script to have the `.postinst` extension.