🔥 Change tests to remove files used by tests
Some tests leave behind files in the project directory, cluttering it and
showing up as new files in every “git status”, which is very annoying.

The ContentFilterTest creates filtered and unfiltered test results, probably
for debugging reasons; the other two tests create a directory to work in, but
don’t clean it up after use, only before use.

The ContentFilterTest was fixed by using the JUnit TemporaryFolder @Rule, which
for each test creates a new temporary directory and will take care of removing
it after the test. The other two tests have to resort to removing the directory
themselves after the tests, because here the temporary directory is already
required in the constructor, and JUnit @Rules are only available during the
actual tests, not during the constructor invocation. A larger change would have
been required, but would have also made the test a bit worse.